Use prompt_run macro for config map and default namespace creation
The `cargo fmt` code formatter makes this actually more lines of code,
but there are fewer if statements in the main routine. So it's kind of a
win.

Borrow checking/memory safety rules make it hard to write a function
that _returns_ our Cmd struct, so we cannot write alternate constructors
for this type. That is a bummer, but it's okay to just put the commands
inline for now. We know it works.