BUGFIX: Only use piped IO if we declare our Cmd writes a file
When developing the argo-init command, we needed to, optionally, allow
our Cmd structs to write their output to files. We have to do this in
Rust since we can't use shell redirection ("> txt.txt") like we do on
the wiki.

In Rust we have designed our runner module with the following syntax
flow:

1. Build a Cmd with args, cwd, any file output
2. Invoke the Cmd via the prompt_run macro, which generates a lot of
   boilerplate code around spawning the command, optionally writing to
   files, matching on expected exit codes.

The bug this PR fixes manifested as the following: we weren't seeing any
terminal output when running terraform commands, which was unexpected.
By default, the rust standard library std::proc::Command struct will
inherit the standard streams from the parent process, e.g. our
interactive terminal. But unfortunately we were assigning piped IO every
time we set the working_dir option, instead of the writes_file option.
Setting the working directory is obviously really important when running
terraform comands, so it manifested there, as opposed to the argo-init
command, because kubectl commands can be run from anywhere, basically,
and we were less concerned with their output.

Docs:
* https://doc.rust-lang.org/std/process/struct.Child.html#method.wait_with_output
* https://doc.rust-lang.org/std/process/struct.Command.html#method.stdout
* https://doc.rust-lang.org/std/process/struct.Command.html#method.spawn