Add commands to namespace_init
We need to add the cluster-info configmap, as well as all the shared
stuff when initializing a namespace

The new prompt_run! macro allows any command to be skipped, optionally,
and the program will continue. So users can read the command that will
be executed and decide if they want it. E.g. we don't HAVE to create the
mars namespace.

Also add a little line separator when prompting before a command
execution.